pyproject.toml
[project]
name = "personal-search-layer"
description = "Local-first unified personal search layer with hybrid retrieval, agentic routing, and evaluation suite."
version = "0.1.0"
authors = [{ name = "Ainesh Mohan", email = "aineshmohan@gmail.com" }]
requires-python = ">=3.12,<3.13"
dependencies = [
	"beautifulsoup4>=4.12",
	"faiss-cpu>=1.8.0",
	"numpy>=1.26",
	"pypdf>=4.2",
	"python-docx>=1.1",
	"sentence-transformers>=2.6",
	"streamlit>=1.32",
]

[project.optional-dependencies]
dev = [
	"pytest>=7.4",
	"ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
personal_search_layer = ["router_policy.json"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
